refactor(input): Y/N 입력 뷰 추상화
Y/N 입력이 반복되어 추상화하였습니다.
